pyproject.toml
[project]
# Project metadata. Available keys are documented at:
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata
name = "ingestables"
description = "IngesTables: A Recipe for Tabular Foundation Models"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [{name = "ingestables authors", email="ingestables-team@google.com"}]
classifiers = [  # List of https://pypi.org/classifiers/
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
]
keywords = []

# pip dependencies of the project
# Installed locally with `pip install -e .`
dependencies = [
    "arff>=0.9",
    "absl-py>=2.1",
    "etils[epath]>=1.12",
    "fiddle>=0.3",
    "pandas>=2.1",
    "scikit-learn>=1.6",
    "tensorboard>=2.18",
    "tensorflow-cpu>=2.18",
    "torch>=2.4",
    "transformers>=4.44",
]

# `version` is automatically set by flit to use `ingestables.__version__`
dynamic = ["version"]

[project.urls]
homepage = "https://github.com/google-research/ingestables"
repository = "https://github.com/google-research/ingestables"
changelog = "https://github.com/google-research/ingestables/blob/main/CHANGELOG.md"
# documentation = ""

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install -e .[dev]`
dev = [
    "pytest",
    "pytest-xdist",
    "pylint>=2.6.0",
    "pyink",
]

[tool.pyink]
# Formatting configuration to follow Google style-guide
line-length = 80
unstable = true
pyink-indentation = 2
pyink-use-majority-quotes = true

[build-system]
# Build system specify which backend is used to build/install the project (flit,
# poetry, setuptools,...). All backends are supported by `pip install`
requires = ["flit_core >=3.8,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.sdist]
# Flit specific options (files to exclude from the PyPI package).
# If using another build backend (setuptools, poetry), you can remove this
# section.
exclude = [
  # Do not release tests files on PyPI
  "**/*_test.py",
]
